requirements_current.txt
aenum==3.1.16
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
boto3==1.42.66
botocore==1.42.66
brotli==1.2.0
CacheControl==0.14.4
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cssselect2==0.9.0
cuda-bindings==12.9.4
cuda-pathfinder==1.4.2
cycler==0.12.1
dataclasses-json==0.6.7
deepgram-sdk==4.8.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.1
ffmpeg-python==0.2.0
filelock==3.25.2
firebase_admin==7.2.0
fonttools==4.62.0
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
google-cloud-aiplatform==1.141.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-firestore==2.24.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-genai==1.66.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.0
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.6.0
hyperframe==6.1.0
identify==2.6.17
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.5.0
langchain==1.2.12
langchain-core==1.2.18
langchain-openai==1.1.11
langgraph==1.1.1
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.16
lxml==6.0.2
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
markdown-katex==202406.1035
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
openai==2.26.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
orjson==3.11.7
ormsgpack==1.12.2
outcome==1.3.0.post0
packaging==26.0
pdf2image==1.17.0
pgvector==0.4.2
pillow==12.1.1
platformdirs==4.9.4
playwright==1.58.0
polars==1.38.1
polars-runtime-32==1.38.1
pre_commit==4.5.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
psycopg2==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycountry==26.2.16
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.11.0
pylatexenc==2.10
pyparsing==3.3.2
pypdf==6.8.0
pyphen==0.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-discovery==1.1.3
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
rank-bm25==0.2.2
redis==7.4.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
schemdraw==0.22
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
starlette==0.52.1
streamlink==8.2.1
stripe==14.4.1
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.1.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.3.0
trio==0.33.0
trio-websocket==0.12.2
triton==3.6.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.21
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
virtualenv==21.2.0
weasyprint==68.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
xxhash==3.6.0
yarl==1.23.0
youtube-transcript-api==1.2.4
yt-dlp==2026.3.3
zopfli==0.4.1
zstandard==0.25.0
